add non-local thread store regression harness
